requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
aiogram>=3.4.0

# Database
aiosqlite>=0.19.0
sqlalchemy[asyncio]>=2.0.0

# Scraping (fallback, optional)
playwright>=1.40.0

# Logging & CLI
loguru>=0.7.0
rich>=13.0.0
tabulate>=0.9.0

# Scheduler (asyncio-native alternative to APScheduler)
apscheduler>=3.10.0
